Handle warnings on queries with no subqueries
Warnings on queries with no subqueries would blow up on the SqlQuery#table method.  The subqueries array is populated from the EXPLAIN output, so is empty for inserts/updates/deletes.

The report view now shows "Table unknown" for such queries, not sure what the best way of picking out the table name is for those cases.